Move `Node` to Node.swift
